restrict certain functions to admins only.
added restart.js

+++ commands/history/clear-message-history.js
@@ -1,4 +1,6 @@
 //imports
+const {isAdmin} = require("../../tools/utils");
+
 // mysql
 const mysql = require("mysql2/promise");
 const db = require("../../config/db");
@@ -15,6 +17,10 @@ const description = "Clears all message history from db";
 
 //main
 const execute = async function (client, message) {
+    if (!isAdmin(message.member)) {
+        await sendMessage("You do not have the authority to perform that function.");
+        return false;
+    }
     try {
         await pool.query("DELETE FROM messages WHERE deleted IS NULL");
     } catch (e) {
